XW-4818 | extract createSVG to utils
