+++ app/initializers/qualaroo.js
@@ -1,13 +1,4 @@
-function createSVG(iconName, className) {
-	const svgNS = 'http://www.w3.org/2000/svg';
-	const svg = document.createElementNS(svgNS, 'svg');
-	const use = document.createElementNS(svgNS, 'use');
-	use.setAttributeNS('http://www.w3.org/1999/xlink', 'xlink:href', `#${iconName}`);
-	svg.setAttribute('role', 'img');
-	svg.classList.add(className);
-	svg.appendChild(use);
-	return svg;
-}
+import createSVG from '../utils/create-svg';
 
 export default {
 	name: 'qualaroo',
