+++ app/styles/module/_qualaroo.scss
@@ -3,6 +3,8 @@ $qualaroo-width-small: 310px;
 $qualaroo-width-large: 350px;
 
 .mobile-wiki {
+	// these are styles for Qualaroo NPS surveys
+	// we customised it to make it nicer
 	div#qual_ol {
 		&.qual_ol {
 			background-color: #fff;
